// src/app/components/offers/models/preferencia-usuario.ts
export enum EstadoVeiculo {
  NOVO = 'NOVO',
  SEMI_NOVO = 'SEMI_NOVO',
  USADO = 'USADO',
}

export enum TipoVeiculo {
  SEDAN = 'SEDAN',
  HATCH = 'SEDAN',
  COUPE = 'SEDAN',
  CONVERTIVEL = 'SEDAN',
  SUV = 'SEDAN',
  CROSSOVER = 'SEDAN',
  MINIVAN = 'SEDAN',
  PICKUP = 'SEDAN',
  CAMINHAO_LEVE = 'SEDAN',
  CAMINHAO_PESADO = 'SEDAN',
  VAN = 'SEDAN',
  MOTOCICLETA = 'SEDAN',
  SCOOTER = 'SEDAN',
  BICICLETA = 'SEDAN',
  VEICULO_ELETRICO = 'SEDAN',
  VEICULO_HIBRIDO = 'SEDAN',
  VEICULO_MILITAR = 'SEDAN',
  VEICULO_DE_EMERGENCIA = 'SEDAN',
  VEICULO_INDUSTRIAL = 'SEDAN',
  MOTORHOME = 'SEDAN',
  TRAILER = 'SEDAN',
  VEICULO_AQUATICO = 'SEDAN',
  VEICULO_OFF_ROAD = 'SEDAN',
  TRATOR = 'SEDAN',
  AERONAVE_LEVE = 'SEDAN',
  AERONAVE = 'SEDAN',
}

export enum Combustivel {
  GASOLINA = 'GASOLINA',
  ETANOL = 'ETANOL',
  DIESEL = 'DIESEL',
  GAS_NATURAL = 'GAS_NATURAL',
  ELETRICO = 'ELETRICO',
}

export enum CaracteristicaVeiculo {
  AR_CONDICIONADO = 'AR_CONDICIONADO',
  DIRECAO_HIDRAULICA = 'DIRECAO_HIDRAULICA',
  CAMBIO_AUTOMATICO = 'CAMBIO_AUTOMATICO',
  BANCOS_DE_COURO = 'BANCOS_DE_COURO',
  CONTROLE_CRUISE = 'CONTROLE_CRUISE',
  SISTEMA_SOM_PREMIUM = 'SISTEMA_SOM_PREMIUM',
  SENSOR_ESTACIONAMENTO = 'SENSOR_ESTACIONAMENTO',
  TETO_SOLAR = 'TETO_SOLAR',
  GPS = 'GPS',
  FAROIS_XENON = 'FAROIS_XENON',
  CAMERA_RE = 'CAMERA_RE',
  ASSISTENTE_FAIXA = 'ASSISTENTE_FAIXA',
  CONTROLE_TRACAO = 'CONTROLE_TRACAO',
  PORTA_MALAS_AUTOMATICO = 'PORTA_MALAS_AUTOMATICO',
  AQUECIMENTO_BANCOS = 'AQUECIMENTO_BANCOS',
  VIDROS_ELETRICOS = 'VIDROS_ELETRICOS',
  SENSOR_CHUVA = 'SENSOR_CHUVA',
  BLUETOOTH = 'BLUETOOTH',
  CONTROLE_VOZ = 'CONTROLE_VOZ',
  AIRBAGS_LATERAIS = 'AIRBAGS_LATERAIS',
  CONTROLE_ESTABILIDADE = 'CONTROLE_ESTABILIDADE',
  RECONHECIMENTO_SINAIS = 'RECONHECIMENTO_SINAIS',
  CARREGADOR_WIRELESS = 'CARREGADOR_WIRELESS',
  MONITOR_PRESSAO_PNEUS = 'MONITOR_PRESSAO_PNEUS',
  ACIONAMENTO_START_STOP = 'ACIONAMENTO_START_STOP',
  PORTAS_AUTOMATICAS = 'PORTAS_AUTOMATICAS',
}

export interface PreferenciaUsuario {
  pesoPreco: number;
  pesoQuilometragem: number;
  pesoTipoVeiculo: number;
  tipoVeiculo: TipoVeiculo[];
  pesoCombustivel: number;
  combustivel: Combustivel[];
  pesoEstadoVeiculo: number;
  estadoVeiculo: EstadoVeiculo[];
  pesoCaracteristicas: number;
  caracteristicas: CaracteristicaVeiculo[];
  distanciaMaxima: number;
  coordenadasUsuario: { latitude: number; longitude: number };
}